DHM - Large changes to the infractucture
I got PapaParse to work. This means I can use a CSV file as the data
source.

I successfully replaced my algorithm with Jake's algorithm (with his
permission of course). I modified his functions to receive/return 2D arrays so that I don't have to mess with strings.

I'm able to print out the instruction set to the screen.

I checked the output against the known successful output (as in, the one
that I used to make the blanket)